Split test in cudart
