+++ cuda-sys/src/lib.rs
@@ -1,6 +1,6 @@
 pub mod vector_types;
 pub mod cuda;
-pub mod cuda_runtime;
+pub mod cudart;
 pub mod cublas;
 
 #[test]
@@ -13,3 +13,4 @@ fn cuda_version() {
 }
 
 mod cuda_tests;
+mod cudart_tests;
